+++ app/src/main/java/com/vishaltelangre/nerdcalci/utils/Suggestion.kt
@@ -1,4 +1,4 @@
-package com.vishaltelangre.nerdcalci.ui.calculator
+package com.vishaltelangre.nerdcalci.utils
 
 enum class SuggestionType {
     VARIABLE, LOCAL_FUNCTION, GLOBAL_FUNCTION, CONSTANT, DYNAMIC_VARIABLE
@@ -6,5 +6,7 @@ enum class SuggestionType {
 
 data class Suggestion(
     val name: String,
-    val type: SuggestionType
+    val type: SuggestionType,
+    val matchIndices: List<Int> = emptyList(),
+    val score: Int = 0
 )